+++ src/main/java/com/myapp/repository/EmployeeRepository.java
@@ -4,12 +4,11 @@ import com.myapp.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee,Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee,Long> {
 
     Optional<Employee> findByEmployeeNameAndEmail(String employeeName, String email);
